backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.26.0
pandas==2.1.1
pillow==10.0.1
scikit-learn==1.3.1
torch==2.1.0
torchvision==0.16.0
matplotlib==3.8.0
seaborn==0.13.0
captum==0.6.0
tensorboard==2.14.1
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.2 